approval_sms

+++ app/Http/Controllers/MenuCategoriesController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\MenuCategoryRequest;
 use App\Collection;
 use App\CollectionMenu;
 use App\EditingMenuCategory;
+use App\User;
 use Auth;
 use Illuminate\Http\Request;
 use App\MenuCategory;
@@ -40,9 +41,7 @@ class MenuCategoriesController extends Controller
                 $categories = $categories->orderby('approved', 'asc')->paginate(20);
             }
 
-        }
-
-        elseif($user->admin == 2) {
+        } elseif ($user->admin == 2) {
 
             $categories = MenuCategory::where('restaurant_id', $user->restaurant_id)->where('deleted', 0)->with('editingMenuCategory');
 
@@ -78,7 +77,7 @@ class MenuCategoriesController extends Controller
         return view('menus.menu_category_create', [
             'restaurant' => $restaurant,
             'user' => $user
-            ]);
+        ]);
     }
 
     /**
@@ -92,9 +91,11 @@ class MenuCategoriesController extends Controller
     {
         $restaurant_id = $request->input('restaurant');
         $user = Auth::user();
+
         if ($user->admin == 2) {
             $restaurant_id = $user->restaurant_id;
         }
+
         $category = new MenuCategory();
         $category->restaurant_id = $restaurant_id;
         $category->name_en = $request->input('name_en');
@@ -105,12 +106,31 @@ class MenuCategoriesController extends Controller
         $name = 'category_' . time() . '.' . $image->getClientOriginalExtension();
         $path = public_path('/images');
         $image->move($path, $name);
-        $category->image =  $name;
-        if($user->admin == 1){
+        $category->image = $name;
+        if ($user->admin == 1) {
             $category->approved = 1;
         }
         $category->save();
         if ($category) {
+
+            if ($user->admin == 2) {
+
+                $restaurant = Restaurant::find($user->restaurant_id);
+
+                $superadmin = User::where('admin', 1)->where('group_id', 1)->first();
+
+                $mobile = $superadmin->mobile_number;
+
+                $provider = $restaurant->name_en;
+
+                $content = $provider . ' has added a new menu category.';
+
+
+                // $url = "https://connectsms.vodafone.com.qa/SMSConnect/SendServlet?application=http_gw209&password=zpr885mi&content=$content&destination=974$mobile&source=97772&mask=Qckly";
+
+                // file($url);
+            }
+
             return redirect('/menu_categories/' . $restaurant_id);
         }
     }
@@ -124,10 +144,26 @@ class MenuCategoriesController extends Controller
     public function approve($id)
     {
         $user = Auth::user();
-        if($user->admin ==1){
+        if ($user->admin == 1) {
+
             MenuCategory::where('id', $id)->update(['approved' => 1]);
+
+            $menuCategory = MenuCategory::find($id);
+
+
+            $provider = User::where('admin', 2)->where('group_id', 2)->where('restaurant_id', $menuCategory->restaurant_id)->first();
+
+            $mobile = $provider->mobile_number;
+
+            $content = 'Your added menu category is approved.';
+
+            // $url = "https://connectsms.vodafone.com.qa/SMSConnect/SendServlet?application=http_gw209&password=zpr885mi&content=$content&destination=974$mobile&source=97772&mask=Qckly";
+
+            // file($url);
+
+
             return redirect()->back();
-        }else{
+        } else {
             return redirect()->back();
         }
     }
@@ -135,10 +171,24 @@ class MenuCategoriesController extends Controller
     public function reject($id)
     {
         $user = Auth::user();
-        if($user->admin ==1){
+        if ($user->admin == 1) {
             MenuCategory::where('id', $id)->update(['approved' => 2]);
+
+            $menuCategory = MenuCategory::find($id);
+
+
+            $provider = User::where('admin', 2)->where('group_id', 2)->where('restaurant_id', $menuCategory->restaurant_id)->first();
+
+            $mobile = $provider->mobile_number;
+
+            $content = 'Your added menu category is rejected.';
+
+            // $url = "https://connectsms.vodafone.com.qa/SMSConnect/SendServlet?application=http_gw209&password=zpr885mi&content=$content&destination=974$mobile&source=97772&mask=Qckly";
+
+            // file($url);
+
             return redirect()->back();
-        }else{
+        } else {
             return redirect()->back();
         }
     }
@@ -159,7 +209,7 @@ class MenuCategoriesController extends Controller
                 return redirect()->back();
             }
         }
-        if($user->admin == 1 && $category->editingMenuCategory !== null){
+        if ($user->admin == 1 && $category->editingMenuCategory !== null) {
             $editingMenuCategory = $category->editingMenuCategory;
             return view('menu_category_edit_approve', [
                 'category' => $category,
@@ -170,7 +220,7 @@ class MenuCategoriesController extends Controller
         return view('menus.menu_category_edit', [
             'category' => $category,
             'user' => $user
-            ]);
+        ]);
     }
 
     /**
@@ -190,23 +240,53 @@ class MenuCategoriesController extends Controller
             if (!$category) {
                 return redirect('/menu_categories');
             }
-            EditingMenuCategory::where('category_id', $category->id)->delete();
+
+            $oldEditingMenuCategory = EditingMenuCategory::where('category_id', $id)->first();
+            if ( $oldEditingMenuCategory) {
+                if ( $oldEditingMenuCategory->image) {
+                    File::delete(public_path('images/' .  $oldEditingMenuCategory->image));
+                }
+
+                EditingMenuCategory::where('category_id', $category->id)->delete();
+            }
+
+
             $editingCategory = new EditingMenuCategory();
             $editingCategory->category_id = $category->id;
             $editingCategory->name_en = $request->input('name_en');
             $editingCategory->description_en = $request->input('description_en');
             $editingCategory->name_ar = $request->input('name_ar');
             $editingCategory->description_ar = $request->input('description_ar');
-            if ($request->hasFile('image')){
+
+            if ($request->hasFile('image')) {
                 $image = $request->file('image');
                 $name = 'category_' . time() . '.' . $image->getClientOriginalExtension();
                 $path = public_path('/images');
                 $image->move($path, $name);
-                $editingCategory->image =  $name;
+                $editingCategory->image = $name;
             }
+
             $editingCategory->save();
-            return redirect('/menu_categories');
-        }elseif ($user->admin == 1){
+
+            $restaurant = Restaurant::find($user->restaurant_id);
+
+            $superadmin = User::where('admin', 1)->where('group_id', 1)->first();
+
+            $mobile = $superadmin->mobile_number;
+
+            $provider = $restaurant->name_en;
+
+            $editingProduct = $category->name_en;
+
+            $content = $provider . ' changed '. $editingProduct . ' data.';
+
+
+            // $url = "https://connectsms.vodafone.com.qa/SMSConnect/SendServlet?application=http_gw209&password=zpr885mi&content=$content&destination=974$mobile&source=97772&mask=Qckly";
+
+            // file($url);
+
+
+        } elseif ($user->admin == 1) {
             $category->name_en = $request->input('name_en');
             $category->description_en = $request->input('description_en');
             $category->name_ar = $request->input('name_ar');
@@ -219,7 +299,7 @@ class MenuCategoriesController extends Controller
                 $name = 'category_' . time() . '.' . $image->getClientOriginalExtension();
                 $path = public_path('/images');
                 $image->move($path, $name);
-                $category->image =  $name;
+                $category->image = $name;
             }
             $category->approved = 1;
             $category->save();
@@ -232,25 +312,45 @@ class MenuCategoriesController extends Controller
     public function editApprove(MenuCategoryRequest $request, $id)
     {
         $user = Auth::user();
-        if($user->admin ==1){
+        if ($user->admin == 1) {
+
             $category = MenuCategory::find($id);
+
+            $editingProduct = $category->name_en;
+
             $editingMenuCategory = EditingMenuCategory::where('category_id', $id)->first();
             $restaurant_id = $request->input('restaurant');
             $category->name_en = $request->input('name_en');
             $category->description_en = $request->input('description_en');
             $category->name_ar = $request->input('name_ar');
             $category->description_ar = $request->input('description_ar');
+
             if ($editingMenuCategory->image) {
                 if ($category->image) {
                     File::delete(public_path('images/' . $category->image));
                 }
-                $category->image =  $editingMenuCategory->image;
+                $category->image = $editingMenuCategory->image;
             }
+
             $category->approved = 1;
             $category->save();
+
+
             EditingMenuCategory::where('category_id', $category->id)->delete();
+
+
+            $provider = User::where('admin', 2)->where('group_id', 2)->where('restaurant_id', $category->restaurant_id)->first();
+
+            $mobile = $provider->mobile_number;
+
+            $content = 'Your changes for ' . $editingProduct . ' have been approved.';
+
+            // $url = "https://connectsms.vodafone.com.qa/SMSConnect/SendServlet?application=http_gw209&password=zpr885mi&content=$content&destination=974$mobile&source=97772&mask=Qckly";
+
+            // file($url);
+
             return redirect('/menu_categories/' . $restaurant_id);
-        }else{
+        } else {
             return redirect()->back();
         }
     }
@@ -258,17 +358,36 @@ class MenuCategoriesController extends Controller
     public function editReject($id)
     {
         $user = Auth::user();
-        if($user->admin ==1){
+        if ($user->admin == 1) {
+
             $editingMenuCategories = EditingMenuCategory::where('category_id', $id)->get();
+
             $category = MenuCategory::find($id);
+            $editingProduct = $category->name_en;
+
             $category_images = [];
-            foreach ($editingMenuCategories as $editingMenuCategory){
+
+            foreach ($editingMenuCategories as $editingMenuCategory) {
                 $category_images[] = public_path('images/' . $editingMenuCategory->image);
             }
+
             File::delete($category_images);
+
             EditingMenuCategory::where('category_id', $id)->delete();
+
+
+            $provider = User::where('admin', 2)->where('group_id', 2)->where('restaurant_id', $category->restaurant_id)->first();
+
+            $mobile = $provider->mobile_number;
+
+            $content = 'Your changes for ' . $editingProduct . ' have been rejected.';
+
+            // $url = "https://connectsms.vodafone.com.qa/SMSConnect/SendServlet?application=http_gw209&password=zpr885mi&content=$content&destination=974$mobile&source=97772&mask=Qckly";
+
+            // file($url);
+
             return redirect('/menu_categories/' . $category->restaurant_id);
-        }else{
+        } else {
             return redirect()->back();
         }
     }
@@ -285,12 +404,12 @@ class MenuCategoriesController extends Controller
         $id = $request->get('id');
         $menuCategory = MenuCategory::where('id', $id)->first();
         if ($user->admin == 2) {
-            if($menuCategory->restaurant_id == $user->restaurant_id){
+            if ($menuCategory->restaurant_id == $user->restaurant_id) {
                 MenuCategory::whereIn('id', $id)->update(['deleted' => 1]);
-            }else{
+            } else {
                 return redirect()->back();
             }
-        }elseif($user->admin == 1){
+        } elseif ($user->admin == 1) {
             MenuCategory::whereIn('id', $id)->update(['deleted' => 1]);
         }
         return redirect('/menu_categories');
